Add context struct and split out functions
A context struct for holding information about a particular tears
session. This includes all the different flags, the rodsEnv, and
the dataObjInp_t for the data object being written to/read from.

This commit reworks the flow of events in this way:

1. iRODS environment information is gathered (via iRODS environment
    file or URI) and stored in the context.
2. dataObjInp_t information is filled out and stored in the context.
3. dataObjInp_t and iRODS environment info are used to select the
    best server via rcGetHostForGet or rcGetHostForPut.
4. Connect to the chosen server.
5. Open or create the specified data object.
6. Enter main loop and stream data as before until finished/error.
7. Close data object, disconnect from server, and exit.

Note: There should be no behavioral changes in this commit.